Add workflow to upload package to Anaconda
Created workflow file to build the package and upload it to Anaconda
when a new tag is pushed. Uses a miniconda3 container to reduce time
needed to run the workflow (fewer package installs). Builds the package,
using 'yq' to generate a sha256 hash for the built .tar.gz, and uses
'conda render' to update the meta.yaml recipe.