Update meta.yaml recipe with version and source url
Updated meta.yaml recipe with the proper dynamic version via load_setup_py_data.
Added a programmatic source url to point to the .tar.gz, for use by the 'yq' package
in the workflow file to generate the sha256 hash. Added updates to reflect changes
made to the publish feature.