Change source url from Github to Pypi
Changed source url from Github to Pypi to resolve hash mismatch caused when sourcing from Github.
The hash of the local tarball built with python-sdist wouldn't match with the hash of source, so
the source has been changed to Pypi and the hash is generated directly from the Pypi tarball.