Distinguish between subscriber message classification and printing
Added separate functions to handle message classification and printing.
Note that the classifier relies only on substring indexing of a message,
which is not a long-term solution to support a diverse mix of messages.
This will hence be updated once a message type schema is determined at
both the publish and subscribe endpoints.