io.py: rework Producer/Consumer used within stream.open()
These features can be split into allowing manual committing of
messages, changing the API slightly for clarity, and making it
easier to hold long-lived references to an open stream.

Manual committing of messages is done by doing mark_done() on
an open stream in read mode. To do this, you can commit the optional
metadata after done processing a message to commit to Kafka.

The long-lived references to a stream is possible via custom
__enter__ and __exit__ methods in the internal Producer and
Consumer classes, so that one can use them without the context
manager interface by explicitly calling .close().

Finally, the API was slightly modified. A new method in the Consumer,
Consumer.read() is used to read messages from a stream in addition
to the iterator API. One can pass in the metadata kwarg here
instead of stream.open() instead which is more explicit and clear.