Further adaptations of the hop publication interface for the robust publisher.
The main changes are the addition of `Producer.write_raw`, and promoting `Producer.pack` to be part of the public interface. In addition, `pack` now processes the headers, with an eye toward it being able to attach new metadata there in the future. A `flush` method is also added to `Producer`.